% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets_VisualisationJH.R
\docType{data}
\name{color_scales_example_data}
\alias{color_scales_example_data}
\title{Example data for discrete color scales}
\format{
A data frame with columns Var_1, Var_2, and class.
}
\usage{
color_scales_example_data
}
\description{
Simulated two-dimensional data with multiple classes.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Example data used to demonstrate discrete color scales.
}\if{html}{\out{</div>}}
}
\keyword{color_scales_example_data}
