% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_plots_ggplot2.R
\name{density_plot_JH}
\alias{density_plot_JH}
\title{Density plot using ggplot2 package, with or without stratification by a factor variable}
\usage{
density_plot_JH(
  data,
  factor = NULL,
  factor_name = NULL,
  color_scale = NULL,
  custom_theme = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{data}{Numeric data for the density plot.}

\item{factor}{Optional factor variable for stratification.}

\item{factor_name}{Optional name of the factor variable.}

\item{color_scale}{Optional hexadecimal color scale for the density(ies).}

\item{custom_theme}{Optional custom ggplot theme.}

\item{na.rm}{Remove the NA values from the density calculation.}
}
\value{
A ggplot2 object of the density plot
}
\description{
Density plot using ggplot2 package, with or without stratification by a factor variable
}
\examples{
data("Pima_data")
Pima_data=Pima_data[complete.cases(Pima_data), ]
density_plot_JH(Pima_data[,"bmi"],
                    color_scale = NULL)
density_plot_JH(Pima_data[,"bmi", drop=FALSE], factor=Pima_data$type,
                    factor_name="Type")
density_plot_JH(Pima_data[,"bmi", drop=FALSE],
                    color_scale = "#0307fc")

}
