% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_themes.R
\name{distinct_scale_bar}
\alias{distinct_scale_bar}
\title{A nice set of colors for plots}
\usage{
distinct_scale_bar()
}
\value{
A nice set of colors for plots
}
\description{
A nice set of colors for plots
}
\examples{
library(ggplot2)
data(color_scales_example_data)
color_scales_example_data=
  color_scales_example_data[color_scales_example_data$class\%in\%c(1:length(distinct_scale_bar())), ]
ggplot(color_scales_example_data) + geom_point(aes(x=Var_1,y=Var_2, color=class)) +
  scale_color_manual("Class", values=distinct_scale_bar())

}
