% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_plots_ggplot2.R
\name{histogram_plot_JH}
\alias{histogram_plot_JH}
\title{Histogram plot using the ggplot2 package}
\usage{
histogram_plot_JH(
  data,
  na.rm = FALSE,
  n_bins = 30,
  plot_density = TRUE,
  color_scale = NULL
)
}
\arguments{
\item{data}{numeric data for the histogram (can be a vector, or a single column data frame).}

\item{na.rm}{remove the NA values from the density calculation? (default: FALSE)}

\item{n_bins}{number of bins for the histogram. (default: 30)}

\item{plot_density}{plot density on top of the histogram? (default: TRUE)}

\item{color_scale}{optional color scale. (default: NULL)}
}
\value{
a ggplot2 object that can be printed
}
\description{
Histogram plot using the ggplot2 package, with the possibility of adding a density line
}
\details{
The histogram is scaled to density so an optional density curve can be overlaid.
}
\examples{
data("Pima_data")
Pima_data=Pima_data[complete.cases(Pima_data), ]
color_scale=c("red", "darkgrey", "navy")

histogram_plot_JH(Pima_data$glucose, na.rm = FALSE, n_bins=30, plot_density=TRUE,
                            color_scale=color_scale)
histogram_plot_JH(Pima_data$glucose, na.rm = FALSE, n_bins=30, plot_density=FALSE)

}
