% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_methods_specific_plots.R
\name{OOB_error_plot}
\alias{OOB_error_plot}
\title{Random forest out-of-bag error plot}
\usage{
OOB_error_plot(randomForest_res, color_scale = NULL)
}
\arguments{
\item{randomForest_res}{randomForest object, resulting from training}

\item{color_scale}{color scale for the plot}
}
\value{
a ggplot2 object of random forest out of bag error plot
}
\description{
Random forest out-of-bag error plot
}
\examples{
data(Pima_data)
Pima_data=Pima_data[complete.cases(Pima_data), ]
X = Pima_data[, names(Pima_data) != "outcome"]
y = Pima_data[, "outcome"]
randomForest_res = randomForest::randomForest(x = X, y = y)
OOB_error_plot(randomForest_res)

}
