% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets_VisualisationJH.R
\docType{data}
\name{Pima_data}
\alias{Pima_data}
\title{Pima diabetes data}
\format{
A data frame.
}
\usage{
Pima_data
}
\description{
Processed Pima diabetes data for examples in the package.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Example data used by plotting functions in this package.
}\if{html}{\out{</div>}}
}
\keyword{Pima_data}
