% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_methods_specific_plots.R
\name{RF_partial_plot}
\alias{RF_partial_plot}
\title{Plot a random forest partial dependence curve}
\usage{
RF_partial_plot(partial_plot_res, color_scale = NULL)
}
\arguments{
\item{partial_plot_res}{Output from \code{randomForest::partialPlot(..., plot = FALSE)}.}

\item{color_scale}{Optional color scale. The first color is used for the curve.}
}
\value{
a ggplot2 object of a random forest partial dependence plot
}
\description{
Plot a random forest partial dependence curve
}
\examples{
data(Pima_data)
Pima_data=Pima_data[complete.cases(Pima_data), ]
X = Pima_data[, names(Pima_data) != "outcome"]
y = Pima_data[, "outcome"]

randomForest_res=randomForest::randomForest(x=X, y=y,
                              ntree=100 ,mtry=3,
                              importance=TRUE, proximity=TRUE)

partial_plot_res = randomForest::partialPlot(
  randomForest_res,
  pred.data = X,
  x.var = "glucose",
  plot = FALSE
)

RF_partial_plot(partial_plot_res)

}
