% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_methods_specific_plots.R
\name{RF_variable_importance_plot}
\alias{RF_variable_importance_plot}
\title{Random forest variable importance plot}
\usage{
RF_variable_importance_plot(randomForest_res, color_scale = NULL)
}
\arguments{
\item{randomForest_res}{randomForest object, resulting from training}

\item{color_scale}{color scale for the plot}
}
\value{
a ggplot2 object of random forest variable importance plot
}
\description{
Random forest variable importance plot
}
\examples{
data(Pima_data)
Pima_data=Pima_data[complete.cases(Pima_data), ]
X = Pima_data[, names(Pima_data) != "outcome"]
y = Pima_data[, "outcome"]
randomForest_res = randomForest::randomForest(x = X, y = y)
RF_variable_importance_plot(randomForest_res)

}
