% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_methods_specific_plots.R
\name{SVM_2_class_decision_boundary}
\alias{SVM_2_class_decision_boundary}
\title{Plot 2D SVM Decision Boundary with Support Vectors}
\usage{
SVM_2_class_decision_boundary(data, svm_res)
}
\arguments{
\item{data}{A data frame containing at least three columns: two numeric features (to be used as axes in the plot)
and a factor column representing the real class of each data point.}

\item{svm_res}{An object of class \code{svm}, returned by the \code{e1071::svm()} function,
representing the trained SVM model.}
}
\value{
A ggplot2 object showing the 2D decision boundary, data points, and support vectors.
}
\description{
This function generates a 2D plot visualizing the decision boundary of
a trained Support Vector Machine (SVM) model.
The plot includes support vectors, real classes of the data points, and the decision regions for each class.
}
\details{
The function first identifies which data points
are support vectors based on the indices provided by \code{svm_res$index}.
It then constructs a grid of values to predict the class labels for each grid point using the trained SVM model,
allowing the decision boundary to be plotted.

The plot displays the decision boundary by using a color gradient to represent different predicted classes.
Data points are overlaid on the decision boundary with a different shape to distinguish support vectors.
The real class of each data point is indicated by color.
}
\examples{
data(Pima_data)
Pima_data=Pima_data[complete.cases(Pima_data), ]
data_decision_boundary=Pima_data[, c("bmi", "age", "outcome")]
decision_boundary_svm_res=e1071::svm(formula=outcome ~ bmi + age, data=data_decision_boundary)
print(SVM_2_class_decision_boundary(data_decision_boundary, decision_boundary_svm_res))

}
