% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_themes.R
\name{theme_publication}
\alias{theme_publication}
\title{A theme for ggplot2}
\usage{
theme_publication(base_size = 14, base_family = "sans")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A theme for ggplot2
}
\description{
A theme for ggplot2
}
\examples{
data(Pima_data)
library(ggplot2)
ggplot(Pima_data) + geom_point(aes(x=bmi,y=glucose)) + ggtitle("Plot title") + theme_publication()

}
